Add missing struct fields and fix compilation errors
diagnosis struct:
- Add p_hosp_diagnosis, p_case_detection[20], case_detection_start_end_yrs[2]
- Add years_btw_case_detection, min_cd_age, min_cd_pack_years, min_cd_symptoms
- Add case_detection_methods[3][20], case_detection_methods_eversmokers[3][20]
- Add case_detection_methods_symptomatic[3][20]
- Increase diagnosis arrays from [8] to [9] to fix array bounds warning

medication struct:
- Add medication_costs[32], medication_utility[32], medication_adherence
- Add ln_h_start_betas_by_class[5][8], ln_h_stop_betas_by_class[5][8]
- Add ln_rr_exac_by_class[5]

model.cpp:
- Fix max() type mismatch by casting MED_CLASS_* enums to int

+++ src/model.cpp
@@ -629,7 +629,7 @@ double update_prevalent_diagnosis(agent *ag)
     {
           if (rand_unif() < input.medication.medication_adherence)
           {
-            (*ag).medication_status= max(MED_CLASS_SABA, (*ag).medication_status);
+            (*ag).medication_status= max((int)MED_CLASS_SABA, (*ag).medication_status);
             medication_LPT(ag);
           }
     }
@@ -638,7 +638,7 @@ double update_prevalent_diagnosis(agent *ag)
     {
           if (rand_unif() < input.medication.medication_adherence)
           {
-            (*ag).medication_status= max(MED_CLASS_LAMA, (*ag).medication_status);
+            (*ag).medication_status= max((int)MED_CLASS_LAMA, (*ag).medication_status);
             medication_LPT(ag);
           }
     }
@@ -701,7 +701,7 @@ double update_prevalent_diagnosis(agent *ag)
       {
         if (rand_unif() < input.medication.medication_adherence)
           {
-               (*ag).medication_status= max(MED_CLASS_SABA, (*ag).medication_status);
+               (*ag).medication_status= max((int)MED_CLASS_SABA, (*ag).medication_status);
                medication_LPT(ag);
           }
       }
@@ -710,7 +710,7 @@ double update_prevalent_diagnosis(agent *ag)
       {
           if (rand_unif() < input.medication.medication_adherence)
           {
-            (*ag).medication_status= max(MED_CLASS_LAMA, (*ag).medication_status);
+            (*ag).medication_status= max((int)MED_CLASS_LAMA, (*ag).medication_status);
             medication_LPT(ag);
           }
       }
@@ -783,7 +783,7 @@ double update_prevalent_diagnosis(agent *ag)
               {
                   if (rand_unif() < input.medication.medication_adherence)
                     {
-                      (*ag).medication_status= max(MED_CLASS_SABA, (*ag).medication_status);
+                      (*ag).medication_status= max((int)MED_CLASS_SABA, (*ag).medication_status);
                       medication_LPT(ag);
                     }
               }
